platform: add platform specific directory traversals
This commit performs the unenviable task of implementing directory
traversal for Windows, Unix-like platforms and Linux (where Linux is
supported via the generic Unix implementation, but also has its own
specialized implementation that uses getdents64 directly).

The purpose of doing this is largely to control our costs more
explicitly, and to fix platform specific bugs. For costs, this largely
boils down to amortizing allocation. That is, instead of allocating
a fresh OsString for every entry, we can now read directory entries
into a previously used entry. Anecdotally, this leads to a 20-30%
performance improvement on listing a single large directory for *both*
Windows and Linux.

As for bugs, these mostly center around long file paths. For example,
on Unix, the typical maximum path length is 4096 bytes. There is no
way to avoid this using std's APIs, so we need to instead roll our own
based on file descriptors. There's also Windows, which has a maximum
file path length of only 260 chars, although this is trickier to fix.

There are basically two downsides to doing things this way:

1. This increases maintenance burden quite a bit, and this change will
   almost certainly introduce bugs on less well tested platforms.
2. It's not clear yet whether we will *also* need to hand-roll our own
   `Metadata` implementation. std provides no way to build the
   `Metadata` type from raw inputs. The key factor here is that std's
   `DirEntry::metadata` method benefits from using `fstatat` internally.
   If we don't re-roll our own `Metadata`, then we'll have to use
   `std::fs::metadata`, which does a normal `stat` call on the file
   path. `fstatat` should generally be faster. On the other hand, std's
   implementation strategy also means that file descriptors aren't
   closed until the last `DirEntry` is dropped, which is a bit sneaky.
   (And is also indicative of a bug in walkdir, since this means its
   "maximum opened fds" feature doesn't actually work.)

Unfortunate, but everything gets sacrificed at the alter of performance.